pyproject.toml
[project]
name = "katha-base"
version = "0.1.0"
description = "Modular storybook system for creating interconnected children's stories"
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "pyyaml>=6.0",
    "openai>=1.0.0",
    "replicate>=0.25.0",
    "requests>=2.31.0",
    "python-dotenv>=1.0.0",
]

[dependency-groups]
dev = [
    "pytest>=7.0.0",
]
